+++ src/compute-engine/boxed-expression/abstract-boxed-expression.ts
@@ -17,7 +17,6 @@ import type {
   PatternMatchOptions,
   Rule,
   RuntimeScope,
-  SemiBoxedExpression,
   Sign,
   SimplifyOptions,
   Substitution,